updating create client control file

+++ api/contollers/clientController.js
@@ -6,6 +6,7 @@ import Supplier from '../model/supplierModel.js';
 import Client from '../model/clientModel.js';
 
 
+
 export const createClient = asyncHandler(async(req,res,next)=>{
 
     // const {error} = luggageValidation(req.body)
@@ -21,57 +22,52 @@ export const createClient = asyncHandler(async(req,res,next)=>{
     //assigning supplier_id and trip_id so as to be able to update the weight field
     
     const supplier_id = supplier._id
-
-    const trip_id = supplier.trip_id
+    // console.log(supplier)
+    const trip_id = supplier.trip._id
 
     //creating new client
     
-    const create_client = await Client.create({...req.body,supplier_name:supplier.name,supplier_id:supplier_id})
+    const create_client = await Client.create({...req.body,supplier:{_id:supplier._id,name:supplier.name}})
    
     //updating supplier number of clients
 
-    const updatedSupplier = await Supplier.findByIdAndUpdate({_id:supplier_id},{$push:{client_name:create_client.name,
-    client_id:create_client.id,client_weight:create_client.weight}},{new:true})
-
-
+    const updatedSupplier = await Supplier.findByIdAndUpdate({_id:supplier_id},{$push:{clients:{name:create_client.name,
+    _id:create_client.id,weight:create_client.weight}}},{new:true})
 
    //updating supplier weight
-   const arr = updatedSupplier.client_weight
-   let new_supplier_weight = 0;
-   for (let i = 0; i < arr.length; i++) {
+    const arr = updatedSupplier.clients
+    let new_supplier_weight = 0;
+    for (let i = 0; i < arr.length; i++) {
+        new_supplier_weight += arr[i].weight;
+    }
 
-    new_supplier_weight += arr[i];
-    
-     
-   }
-
-
-    const updatedSupplierWeight = await Supplier.findByIdAndUpdate({_id:supplier_id},{$set:{number_clients:arr.length,
+    await Supplier.findByIdAndUpdate({_id:supplier_id},{$set:{number_clients:arr.length,
     weight:new_supplier_weight,}},{new:true})
 
-    console.log(updatedSupplierWeight)
+    // console.log(updatedSupplierWeight)
+    const trip = await Trip.findOne({'suppliers._id':supplier_id})
+    if(!trip) return res.status(400).send('supplier with id doesnt exist')
+    // console.log(trip)
 
-
-
+    const supplier_trip = await Trip.findOneAndUpdate({'suppliers._id':supplier_id},{$set:{'suppliers.$':{name:supplier.name,_id:supplier_id,weight:new_supplier_weight}}},{new:true})
+    if(!supplier_trip) res.status(400).json({"message":"failed to update supplier trip"})
+    console.log(supplier_trip)
 
     //updating trip weight
 
-    
-    const updatedTrip = await Trip.findByIdAndUpdate({_id:trip_id},{$push:{supplier_weight:new_supplier_weight,}},{new:true})
 
-    // const  = updatedSupplier.client_weight
+    console.log(supplier_trip.suppliers.length)
     let new_trip_weight = 0;
-    for (let i = 0; i < updatedSupplier.client_weight.length; i++) {
- 
-     new_trip_weight += arr[i];
-     
+    for (let i = 0; i < supplier_trip.suppliers.length; i++) {
+     new_trip_weight += supplier_trip.suppliers[i].weight;
     }
 
+    console.log(new_trip_weight)
 
     await Trip.findByIdAndUpdate({_id:trip_id},{$set:{weight:new_trip_weight}},{new:true})
-    console.log()
 
-    //sending response to the client
+    // sending response to the client
+    
     res.status(200).send(create_client)
    
 })
@@ -109,10 +105,54 @@ export const getClient = asyncHandler(async(req,res)=>{
 export const  deleteClient = async(req,res,next)=>{
 
     try{
+
+        const client = await Client.findById(req.params.id)
+        if(!client) return res.status(400).send('client with id doesnt exist')
+        const supplier_id = client.supplier_id
+
+       
  
-
+        const supplier = await Supplier.findById(supplier_id)
+        if(!supplier) return res.status(400).send('supplier with id doesnt exist')
+        console.log(supplier)
+    
+        // assigning supplier_id and trip_id so as to be able to update the weight field
+        
+        const trip_id = supplier.trip_id
+    
+       
+    
+        const updatedSupplier = await Supplier.findByIdAndUpdate({_id:supplier_id},{$pull:{client_name:client.name,
+        client_id:client.id,client_weight:client.weight}},{new:true})
+    
+       //updating supplier weight
+       const arr = updatedSupplier.client_weight
+       let new_supplier_weight = 0;
+       for (let i = 0; i < arr.length; i++) {
+        new_supplier_weight += arr[i];
+       }
+    
+        const updatedSupplierWeight = await Supplier.findByIdAndUpdate({_id:supplier_id},{$set:{number_clients:arr.length,
+        weight:new_supplier_weight,}},{new:true})
+    
+        console.log(updatedSupplierWeight)
+    
+        //updating trip weight
+    
+        const updatedTrip = await Trip.findByIdAndUpdate({_id:trip_id},{$set:{supplier_weight:new_supplier_weight,}},{new:true})
+    
+        // const  = updatedSupplier.client_weight
+        let new_trip_weight = 0;
+        for (let i = 0; i < updatedSupplier.client_weight.length; i++) {
+         new_trip_weight += arr[i];
+        }
+    
+        await Trip.findByIdAndUpdate({_id:trip_id},{$set:{weight:new_trip_weight}},{new:true})
+        console.log()
+    
+        // sending response to the client
+        
         await Client.findByIdAndDelete(req.params.id)
-
         res.status(200).json({"success":true})
 
     }catch(error){
